#22 - add unit tests
