+++ client/src/app/checkout/page.tsx
@@ -1,10 +1,9 @@
 'use client';
 
 import React from 'react';
-import InputField from '@/src/components/InputField';
+
 import {
   Box,
-  Button,
   Card,
   CardContent,
   CardHeader,
@@ -14,66 +13,9 @@ import {
 } from '@mui/material';
 import Image from 'next/image';
 import VR from '@/src/images/vr1.png';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { CheckoutFormSchema } from '@/src/utils/types';
-import { checkoutSchema } from '@/src/utils/validationSchemat';
-
-const inputs: {
-  fieldName:
-    | 'firstName'
-    | 'lastName'
-    | 'email'
-    | 'phoneNumber'
-    | 'country'
-    | 'city'
-    | 'street'
-    | 'postalCode';
-  labelName: string;
-}[] = [
-  {
-    fieldName: 'firstName',
-    labelName: 'First Name',
-  },
-  {
-    fieldName: 'lastName',
-    labelName: 'Last Name',
-  },
-  {
-    fieldName: 'email',
-    labelName: 'Email address',
-  },
-  {
-    fieldName: 'phoneNumber',
-    labelName: 'Phone number',
-  },
-  {
-    fieldName: 'country',
-    labelName: 'Country',
-  },
-  {
-    fieldName: 'city',
-    labelName: 'City',
-  },
-  {
-    fieldName: 'street',
-    labelName: 'Street',
-  },
-  {
-    fieldName: 'postalCode',
-    labelName: 'Postal dode',
-  },
-];
+import CheckoutForm from '@/src/components/CheckoutForm';
 
 const CheckoutPage = () => {
-  const { control, handleSubmit } = useForm<CheckoutFormSchema>({
-    resolver: zodResolver(checkoutSchema),
-  });
-
-  const onSubmit: SubmitHandler<CheckoutFormSchema> = (data) => {
-    console.log(data);
-  };
-
   return (
     <Box>
       <Typography variant="h1" fontSize="25px">
@@ -81,46 +23,7 @@ const CheckoutPage = () => {
       </Typography>
       <Grid container spacing={8}>
         <Grid item xl={8}>
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <Grid container marginTop="20px" rowSpacing={5} columnSpacing={10}>
-              {inputs.map(({ fieldName, labelName }) => (
-                <Grid key={fieldName} item md={6}>
-                  <Controller
-                    name={fieldName}
-                    control={control}
-                    render={({ field, fieldState: { error } }) => (
-                      <InputField
-                        labelName={labelName}
-                        helperText={error ? error.message : null}
-                        {...field}
-                      />
-                    )}
-                  />
-                </Grid>
-              ))}
-              <Grid
-                item
-                sm={12}
-                sx={{ display: 'flex', justifyContent: 'center' }}
-              >
-                <Button
-                  type="submit"
-                  sx={{
-                    color: 'white',
-                    padding: '6px 50px',
-                    border: '1px solid white',
-                    background: 'rgba(255, 255, 255, 0.2)',
-                    borderRadius: '0',
-                    marginTop: '10px',
-                    fontSize: '18px',
-                    textTransform: 'capitalize',
-                  }}
-                >
-                  Pay
-                </Button>
-              </Grid>
-            </Grid>
-          </form>
+         <CheckoutForm />
         </Grid>
         <Grid item xl={4}>
           <Card
