+++ client/src/components/CheckoutForm/inputs.ts
@@ -1,0 +1,45 @@
+export const checkoutInputs: {
+  fieldName:
+    | 'firstName'
+    | 'lastName'
+    | 'email'
+    | 'phoneNumber'
+    | 'country'
+    | 'city'
+    | 'street'
+    | 'postalCode';
+  labelName: string;
+}[] = [
+  {
+    fieldName: 'firstName',
+    labelName: 'First Name',
+  },
+  {
+    fieldName: 'lastName',
+    labelName: 'Last Name',
+  },
+  {
+    fieldName: 'email',
+    labelName: 'Email address',
+  },
+  {
+    fieldName: 'phoneNumber',
+    labelName: 'Phone number',
+  },
+  {
+    fieldName: 'country',
+    labelName: 'Country',
+  },
+  {
+    fieldName: 'city',
+    labelName: 'City',
+  },
+  {
+    fieldName: 'street',
+    labelName: 'Street',
+  },
+  {
+    fieldName: 'postalCode',
+    labelName: 'Postal dode',
+  },
+];
